Indicator service y variable service modificados

+++ src/modules/spadmin/services/variable.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 import { Variable } from '../entities/variable.entity';
+import { Status } from '../../users/entities/status.entity';
 import { CreateVariableDto, UpdateVariableDto } from '../dtos/variable.dto';
 
 @Injectable()
@@ -10,6 +11,7 @@ export class VariableService {
   constructor(
     @InjectRepository(Variable)
     private variableRepo: Repository<Variable>,
+    @InjectRepository(Status) private statusRepo: Repository<Status>,
   ) {}
 
   findAll() {
@@ -19,7 +21,7 @@ export class VariableService {
   findOne(id: number) {
     const user = this.variableRepo.findOne({
       where: { id: id },
-      // relations: ['user'],
+      relations: ['status'],
     });
     if (!user) {
       throw new NotFoundException(`Dni Type #${id} not found`);
@@ -27,18 +29,34 @@ export class VariableService {
     return user;
   }
 
-  create(data: CreateVariableDto) {
+  async create(data: CreateVariableDto) {
     const newVariable = this.variableRepo.create(data);
+    if (data.statusId) {
+      const status = await this.statusRepo.findOne({
+        where: { id: data.statusId },
+      });
+      newVariable.status = status;
+    }
     return this.variableRepo.save(newVariable);
   }
 
   async update(id: number, changes: UpdateVariableDto) {
     const variable = await this.variableRepo.findOne({ where: { id: id } });
+    if (changes.statusId) {
+      const status = await this.statusRepo.findOne({
+        where: { id: changes.statusId },
+      });
+      variable.status = status;
+    }
     this.variableRepo.merge(variable, changes);
     return this.variableRepo.save(variable);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const variable = await this.findOne(id);
+    if (!variable) {
+      throw new NotFoundException(`Variable #${id} not found`);
+    }
     return this.variableRepo.delete(id);
   }
 }
